+++ packages/base/src/form-element.ts
@@ -15,14 +15,13 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {addHasRemoveClass, BaseElement, CustomEventListener, EventType, observer, SpecificEventListener} from './base-element.js';
+import {addHasRemoveClass, BaseElement, CustomEventListener, EventType, SpecificEventListener} from './base-element.js';
 
 export {
   addHasRemoveClass,
   BaseElement,
   CustomEventListener,
   EventType,
-  observer,
   SpecificEventListener
 };