+++ packages/list/src/mwc-list-base.ts
@@ -15,8 +15,9 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-import {BaseElement, observer} from '@material/mwc-base/base-element.js';
-import {deepActiveElementPath, doesElementContainFocus, findAssignedElement, isNodeElement} from '@material/mwc-base/utils';
+import {BaseElement} from '@material/mwc-base/base-element.js';
+import {observer} from '@material/mwc-base/observer.js';
+import {deepActiveElementPath, doesElementContainFocus, findAssignedElement, isNodeElement} from '@material/mwc-base/utils.js';
 import {html, property, query} from 'lit-element';
 import {ifDefined} from 'lit-html/directives/if-defined.js';